+++ HISTORY.rst
@@ -11,6 +11,7 @@ History
     * Card #171: improve and fix Order reshoot transition do update payout values on the old assignment and copy old values to the new assignment (rudaporto).
     * Card #167: improve Assignment workflow transition retract_rejection to move from Awaiting Assets to In QA without resubmit (rudaporto).
     * Card #41: update Assignment perm_reject transition and subscriber since it will be now called only from the Order workflow (rudaporto).
+    * Remove payout_currency from Order transitions new_shoot, perm_reject and reshoot (rudaporto).
 
 2.1.1 (2017-03-29)
 ------------------